1. 改了下删除条目的接口 2. 当删除的条目不存在的时候, 返回error=2

+++ wiki/src/main/java/tss2/wiki/control/ContentController.java
@@ -142,6 +142,9 @@ public class ContentController {
             return new CommonResult(1, "Authentication Failed");
         }
         WikiRecord wikiRecord = new WikiRecord(title);
+        if (wikiRecord.getContent() == null) {
+            return new CommonResult(2, "Entry Not Exist");
+        }
         wikiRecord.delete();
         return new CommonResult(0);
     }
